First Copy of Working Module Uploaded
The content that is not part of the actual boa class is just me testing
out the module whenever I make changes to see if it works or make sure I
don't break anything. The string variable data is a list scraped off of
another site that I was using for testing.