Uploaded newBoa.py, a new version of module
Probably going to eventually make this the main version of the module
and keep the old one for reference or something. Now my idea for this
module is for you to input a string that contains http response data and
dynamically parses it in a structure that contains information of all of
the tags in a document, whether their nested or not, what there nested
in, and some other random stuff. Would be cool if it could interpret
javascript that would normally change a page. Also want to use this
library to make graphical tool for analyzing a website for communication
by finding the patterns of dynamic content generation